[Refactor] 코딩 테스트 페이지에서 훅, 상수 파일 분리
